WIP: Add "mindiff_overrides" key to config to allow per-client, useragent based raised minimum difficulty (#13)
This implements the new `mindiff_overrides` config key.  The key is documented as follows:

---
```
"mindiff_overrides" : Optional JSON object (dictionary) of useragent, mindiff
pairs, e.g: `"mindiff_overrides" : { "nicehash" : 500000, "foohash" : 750000 }`
This is a mechanism to support a raised minimum difficulty for clients
connecting to the pool based on their "useragent" string (that they normally
send via `mining.subscribe` on initial connect).  The "useragent" specified
here is a prefix match (case insensitive).  If a client's useragent matches the
specified prefix, then it will have its minimum and starting difficulty set to
the override value.  Note: The override values specified here cannot be lower
than the pool's "mindiff" setting -- they must always be higher (similarly this
value must also be below "maxdiff").  This mechanism was initially designed so
that NiceHash clients can connect to a pool and get a special-cased raised
difficulty  (since NiceHash refuses to operate if share difficulty is < 500000
as of the time of this writing).
```

---

This feature re-uses the same machinery that was in place for the `mining.suggest_difficulty` stratum protocol method.  Basically, if a client's useragent string matches an override (prefix based match, case insensitive), then its minimum difficulty and starting difficulty will be set to the override, as if it had sent a `mining.suggest_difficulty` command to the server.

Note this has been tested by @tfaoro preliminarily, and appears to work as intended. The changes are minimally invasive, so it is safe to merge now.