Rename ckpool, ckdb, etc -> asicseer-pool, asicseec-db, etc (#15)
Fixes #14 .

As per the discussion in #14, renames:

- `ckpool` -> `asicseer-pool`
- `ckdb` -> `asicseer-db`
- `ckpmsg` -> `asicseer-pmsg`

Also renames the socket dir from  `/tmp/ckpool` -> `/tmp/asicseer-pool` (or `/tmp/asicseer-passthrough`, etc, depending on mode).

The log data files are unchanged, but the logfile for debug/log output is now `logs/asicseer-pool.log`.

Also the `.conf` files were renamed to `asicseer-*`, and the program looks for those default `asiceseer-*` conf files now by default if no `-c` is specified on the CLI.

Tested and reviewed -- works.

---

Also in this commit: Refuse to proceed it bitcoind does not support `isscript` on
`validateaddress`, since we really need this facility for correct operation. (Shouldn't
be a problem on modern bitcoind's).