Fix: -D mode was broken.  Now it is no longer broken.
-D mode would fork() too late, thus losing some of the opened threads
(including the logger thread). Likely ck didn't know how fork() actually
works when he implemented this facility.

The fix is to fork ASAP if -D is on the CLI, before starting any
threads.

Note that this means that if there's an error later the calling process
won't know about it -- but that's typical for daemonizing programs (even
bitcoind has this problem).

Now, -D correctly outputs log files, etc, as intended.

Bumped version to 1.0.3.