Renamed British "authorise" -> US "authorize", app-wide
This includes the ckdb stuff, etc.  Internally the app was refering to
"mining.authorize" as "authorise" and the inconsistency bothered me.
Since the stratum command is "minint.authorize" and it is standard and
cannot be changed, we opted to go with the US spelling of the word app-wide
for consistency.  That and also I am American.

Bumped app version to 1.1.1.