Improved ZMQ reliability
ZMQ is very unreliable. If the bitcoind goes away and comes back, the
socket can be dead without a way to test that it died.

So, we added logic to detect when it appears the ZMQ sockets are dead,
and to recreate them in that case.

Also in this commit we verify/validate the bitcoind configuration versus
what we are expecting by hitting the getzmqnotifications RPC on bitcoind.

For tcp:// endpoints, we simply verify that the protocol and port matches
(since bitcoind might be binding to 0.0.0.0 or some other interface).
For ipc:// endpoints, we verify that we can access the file corresponding
to the socket.

Also we added some better thread safety to updating the sdata->lasthash
data.

Bumped version to 1.4.3