Fixed slowness of json-rpc to bitcoind
JSON rpc calls to bitcoind couldn't cope with extremely large responses
(as you would get from gbt with a full mempool).  The reason is it was
scanning the input constantly for newlines, and not using the
Content-Length: header field.

It has not been fixed to use the content-length field from the HTTP
response to more intelligently process responses.  The response now is
no longer constantly being scanned for newlines as it is read, but
instead it is read as a big blob of content-length bytes.

This is much faster and saves on CPU overhead when doing RPC to
bitcoind.

Also in this commit:

Bumped version to 1.4.7.